export formatTextSegmentBeforeSelectionMarker
