+++ packages/roosterjs-content-model-plugins/lib/autoFormat/link/createLink.ts
@@ -1,5 +1,5 @@
 import { addLink } from 'roosterjs-content-model-dom';
-import { formatTextSegmentBeforeSelectionMarker } from '../../pluginUtils/formatTextSegmentBeforeSelectionMarker';
+import { formatTextSegmentBeforeSelectionMarker } from 'roosterjs-content-model-api';
 import { matchLink } from 'roosterjs-content-model-api';
 import type { IEditor, LinkData } from 'roosterjs-content-model-types';
