+++ packages/roosterjs-content-model-plugins/test/autoFormat/AutoFormatPluginTest.ts
@@ -1,5 +1,5 @@
 import * as createLink from '../../lib/autoFormat/link/createLink';
-import * as formatTextSegmentBeforeSelectionMarker from '../../lib/pluginUtils/formatTextSegmentBeforeSelectionMarker';
+import * as formatTextSegmentBeforeSelectionMarker from 'roosterjs-content-model-api/lib/publicApi/utils/formatTextSegmentBeforeSelectionMarker';
 import * as unlink from '../../lib/autoFormat/link/unlink';
 import { AutoFormatOptions, AutoFormatPlugin } from '../../lib/autoFormat/AutoFormatPlugin';
 import { createLinkAfterSpace } from '../../lib/autoFormat/link/createLinkAfterSpace';
