+++ packages/roosterjs-content-model-plugins/test/autoFormat/hyphen/transformHyphenTest.ts
@@ -1,5 +1,7 @@
+import { formatTextSegmentBeforeSelectionMarker } from 'roosterjs-content-model-api';
 import { transformHyphen } from '../../../lib/autoFormat/hyphen/transformHyphen';
 import {
+    ContentModelDocument,
     ContentModelParagraph,
     ContentModelText,
     FormatContentModelContext,
@@ -139,5 +141,306 @@ describe('transformHyphen', () => {
             format: {},
         };
         runTest(segment, paragraph, { canUndoByBackspace: true } as any, false);
+    });
+});
+
+describe('formatTextSegmentBeforeSelectionMarker - transformHyphen', () => {
+    function runTest(
+        input: ContentModelDocument,
+        expectedModel: ContentModelDocument,
+        expectedResult: boolean
+    ) {
+        const formatWithContentModelSpy = jasmine
+            .createSpy('formatWithContentModel')
+            .and.callFake((callback, options) => {
+                const result = callback(input, {
+                    newEntities: [],
+                    deletedEntities: [],
+                    newImages: [],
+                    canUndoByBackspace: true,
+                });
+                expect(result).toBe(expectedResult);
+            });
+
+        formatTextSegmentBeforeSelectionMarker(
+            {
+                focus: () => {},
+                formatContentModel: formatWithContentModelSpy,
+            } as any,
+            (_model, previousSegment, paragraph, _markerFormat, context) => {
+                return transformHyphen(previousSegment, paragraph, context);
+            }
+        );
+
+        expect(formatWithContentModelSpy).toHaveBeenCalled();
+        expect(input).toEqual(expectedModel);
+    }
+
+    it('No hyphen', () => {
+        const input: ContentModelDocument = {
+            blockGroupType: 'Document',
+            blocks: [
+                {
+                    blockType: 'Paragraph',
+                    segments: [
+                        {
+                            segmentType: 'Text',
+                            text: 'test',
+                            format: {},
+                        },
+                        {
+                            segmentType: 'SelectionMarker',
+                            isSelected: true,
+                            format: {},
+                        },
+                    ],
+                    format: {},
+                },
+            ],
+            format: {},
+        };
+        runTest(input, input, false);
+    });
+
+    it('with hyphen', () => {
+        const text = 'test--test';
+        spyOn(text, 'split').and.returnValue(['test--test ']);
+        const input: ContentModelDocument = {
+            blockGroupType: 'Document',
+            blocks: [
+                {
+                    blockType: 'Paragraph',
+                    segments: [
+                        {
+                            segmentType: 'Text',
+                            text: text,
+                            format: {},
+                        },
+                        {
+                            segmentType: 'SelectionMarker',
+                            isSelected: true,
+                            format: {},
+                        },
+                    ],
+                    format: {},
+                },
+            ],
+            format: {},
+        };
+
+        const expected: ContentModelDocument = {
+            blockGroupType: 'Document',
+            blocks: [
+                {
+                    blockType: 'Paragraph',
+                    segments: [
+                        {
+                            segmentType: 'Text',
+                            text: 'test—tes',
+                            format: {},
+                            isSelected: undefined,
+                        },
+                        {
+                            segmentType: 'Text',
+                            text: 't',
+                            format: {},
+                            isSelected: undefined,
+                        },
+
+                        {
+                            segmentType: 'SelectionMarker',
+                            isSelected: true,
+                            format: {},
+                        },
+                    ],
+                    format: {},
+                },
+            ],
+            format: {},
+        };
+        runTest(input, expected, true);
+    });
+
+    it('with hyphen and left space', () => {
+        const input: ContentModelDocument = {
+            blockGroupType: 'Document',
+            blocks: [
+                {
+                    blockType: 'Paragraph',
+                    segments: [
+                        {
+                            segmentType: 'Text',
+                            text: 'test-- test',
+                            format: {},
+                        },
+                        {
+                            segmentType: 'SelectionMarker',
+                            isSelected: true,
+                            format: {},
+                        },
+                    ],
+                    format: {},
+                },
+            ],
+            format: {},
+        };
+
+        runTest(input, input, false);
+    });
+
+    it('with hyphen and left space', () => {
+        const input: ContentModelDocument = {
+            blockGroupType: 'Document',
+            blocks: [
+                {
+                    blockType: 'Paragraph',
+                    segments: [
+                        {
+                            segmentType: 'Text',
+                            text: 'test --test',
+                            format: {},
+                        },
+                        {
+                            segmentType: 'SelectionMarker',
+                            isSelected: true,
+                            format: {},
+                        },
+                    ],
+                    format: {},
+                },
+            ],
+            format: {},
+        };
+
+        runTest(input, input, false);
+    });
+
+    it('with hyphen between spaces', () => {
+        const text = 'test -- test';
+        spyOn(text, 'split').and.returnValue(['test', '--', 'test']);
+        const input: ContentModelDocument = {
+            blockGroupType: 'Document',
+            blocks: [
+                {
+                    blockType: 'Paragraph',
+                    segments: [
+                        {
+                            segmentType: 'Text',
+                            text: 'test -- test',
+                            format: {},
+                        },
+                        {
+                            segmentType: 'SelectionMarker',
+                            isSelected: true,
+                            format: {},
+                        },
+                    ],
+                    format: {},
+                },
+            ],
+            format: {},
+        };
+
+        const expected: ContentModelDocument = {
+            blockGroupType: 'Document',
+            blocks: [
+                {
+                    blockType: 'Paragraph',
+                    segments: [
+                        {
+                            segmentType: 'Text',
+                            text: 'test ',
+                            format: {},
+                            isSelected: undefined,
+                        },
+                        {
+                            segmentType: 'Text',
+                            text: '—',
+                            format: {},
+                            isSelected: undefined,
+                        },
+                        {
+                            segmentType: 'Text',
+                            text: ' test',
+                            format: {},
+                            isSelected: undefined,
+                        },
+                        {
+                            segmentType: 'SelectionMarker',
+                            isSelected: true,
+                            format: {},
+                        },
+                    ],
+                    format: {},
+                },
+            ],
+            format: {},
+        };
+        runTest(input, expected, true);
+    });
+
+    it('with hyphen and multiple words', () => {
+        const text = 'testing test--test';
+        spyOn(text, 'split').and.returnValue(['testing', 'test--test ']);
+        const input: ContentModelDocument = {
+            blockGroupType: 'Document',
+            blocks: [
+                {
+                    blockType: 'Paragraph',
+                    segments: [
+                        {
+                            segmentType: 'Text',
+                            text: text,
+                            format: {},
+                        },
+                        {
+                            segmentType: 'SelectionMarker',
+                            isSelected: true,
+                            format: {},
+                        },
+                    ],
+                    format: {},
+                },
+            ],
+            format: {},
+        };
+
+        const expected: ContentModelDocument = {
+            blockGroupType: 'Document',
+            blocks: [
+                {
+                    blockType: 'Paragraph',
+                    segments: [
+                        {
+                            segmentType: 'Text',
+                            text: 'testing ',
+                            format: {},
+                            isSelected: undefined,
+                        },
+                        {
+                            segmentType: 'Text',
+                            text: 'test—tes',
+                            format: {},
+                            isSelected: undefined,
+                        },
+                        {
+                            segmentType: 'Text',
+                            text: 't',
+                            format: {},
+                            isSelected: undefined,
+                        },
+
+                        {
+                            segmentType: 'SelectionMarker',
+                            isSelected: true,
+                            format: {},
+                        },
+                    ],
+                    format: {},
+                },
+            ],
+            format: {},
+        };
+        runTest(input, expected, true);
     });
 });
